+++ nuxt.config.js
@@ -1,8 +1,8 @@
-const environment = process.env.NODE_ENV
-const envSet = require(`./env.${environment}.js`)
+// const environment = process.env.NODE_ENV
+// const envSet = require(`./env.${environment}.js`)
 
 export default {
-  env: envSet,
+  // env: envSet,
 
   // Global page headers: https://go.nuxtjs.dev/config-head
   head: {
@@ -39,21 +39,19 @@ export default {
   modules: [
     // https://go.nuxtjs.dev/axios
     "@nuxtjs/axios",
-    './env.development.js',
-
-
+    // './env.development.js',
   ],
 
   // Axios module configuration: https://go.nuxtjs.dev/config-axios
   axios: {
     // Workaround to avoid enforcing hard-coded localhost:3000: https://github.com/nuxt-community/axios-module/issues/308
     proxy: true,
-    // baseURL: 'http://localhost:3000',
+    baseURL: 'http://localhost:5000',
   },
 
-  proxy: {
-    "/api": envSet.apiBaseUrl
-  },
+  // proxy: {
+  //   "/api": envSet.apiBaseUrl
+  // },
   // Build Configuration: https://go.nuxtjs.dev/config-build
   build: {},
 
@@ -87,4 +85,5 @@ export default {
       baseURL: process.env.apiBaseUrl,
     },
   },
+
 }
