+++ package.json
@@ -13,16 +13,13 @@
     "lintfix": "prettier --write --list-different . && npm run lint:js -- --fix"
   },
   "dependencies": {
-    "@nuxtjs/auth-next": "5.0.0-1667386184.dfbbb54",
+    "@nuxtjs/auth-next": "^5.0.0-1667386184.dfbbb54",
     "@nuxtjs/axios": "^5.13.6",
     "bcrypt": "^5.1.0",
     "core-js": "^3.25.3",
-    "cors": "^2.8.5",
-    "cross-env": "^7.0.3",
     "express": "^4.18.2",
-    "mustache-express": "^1.3.2",
     "n": "^9.0.1",
-    "nuxt": "^3.3.2",
+    "nuxt": "^2.15.8",
     "passport": "^0.6.0",
     "stable": "^0.1.8",
     "vue": "^2.7.10",
