+++ project/package.json
@@ -21,14 +21,18 @@
   "dependencies": {
     "@nuxtjs/auth-next": "5.0.0-1667386184.dfbbb54",
     "bcrypt": "^5.1.0",
+    "connect-flash": "^0.1.1",
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "express-session": "^1.17.3",
     "express-validator": "^6.15.0",
     "mustache-express": "^1.3.2",
     "mysql2": "^3.2.0",
     "nodemailer": "^6.9.1",
+    "passport": "^0.6.0",
+    "passport-local": "^1.0.0",
     "sequelize": "^6.29.3"
   },
   "devDependencies": {
