+++ project/seeders/20230327133858-test-users.js
@@ -1,0 +1,21 @@
+'use strict';
+const bcrypt = require('bcrypt');
+
+module.exports = {
+  up: (queryInterface, Sequelize) => {
+    const now = new Date();
+    return queryInterface.bulkInsert('Users', [
+      {
+        name: '太郎',
+        email: 'taro@example.com',
+        password: bcrypt.hashSync('secret', bcrypt.genSaltSync(8)),
+        createdAt: new Date(),
+        updatedAt: new Date()
+      },
+    ], {});
+  },
+
+  down: (queryInterface, Sequelize) => {
+    return queryInterface.bulkDelete('Users', null, {});
+  }
+};
